feature:created-button
Created a new button UI element to enable users to Create a new Account, but the logic has not yet been implemented. The button was designed with the user experience in mind, and it fits seamlessly with the existing UI design. Although the button does not currently have any logic, its implementation is expected to provide significant value to users in future releases.